requirements.txt
hermes-python
toml
python-snap7
